@@ -1,16 +1,15 @@
 package com.archerprop.peja.repositories;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import com.archerprop.peja.entity.Usuarios;
 
-@Repository
 public interface UsuarioRepositorio extends JpaRepository<Usuarios, Long> {
 
-    public abstract Usuarios findByCorreo(String correo);
+    Optional<Usuarios> findByCorreo(String correo);
 
     public abstract List<Usuarios> findByNombre(String nombre);
 
